requirements.txt
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
DateTime==5.5
idna==3.7
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schedule==1.2.2
soupsieve==2.5
telebot==0.0.5
urllib3==2.2.1
zope.interface==6.4.post2
